api/requirements.txt
fastapi
python-multipart
pillow
numpy
httpx
tflite-runtime
